move charm++ to prod-env, fix cuda library linking
Created symlinks frm the cudaToolkit.lib/lib to cudaToolkit.out/lib64 to satisfy
the expectations of NAMD2. This could be worth a mention to the maintainers; it's not
clear how often this kind of thing must be done, or how many things really depend
on libcudart.so vs. the other cuda libraries.

Signed-off-by: Eric Coulter <jecoulte@iu.edu>